WIP idea - but waiting on feedback
